bugfix - dont copy formats of previous row
Using LO.ListRows.Add, it copies the formatting of the previous
row. If a row above has a `=[foo]` value, it will be formatted as text
and result in error if the next row's item should not be formatted
as text. For example R21.00 ended up being
'$21.00' as string instead of the value 21